Slight reorganization
