+++ jingle/src/modeling/bmc/machine/cpu/concrete.rs
@@ -1,0 +1,62 @@
+use z3::Context;
+use z3::ast::BV;
+use jingle_sleigh::VarNode;
+use crate::modeling::bmc::machine::cpu::symbolic::SymbolicPcodeAddress;
+
+
+pub type PcodeMachineAddress = u64;
+pub type PcodeOffset = u8;
+#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
+pub struct ConcretePcodeAddress {
+    pub(crate) machine: PcodeMachineAddress,
+    pub(crate) pcode: PcodeOffset,
+}
+
+impl ConcretePcodeAddress {
+    pub fn next_pcode(&self) -> Self {
+        self.add_pcode_offset(1)
+    }
+
+    pub fn machine(&self) -> PcodeMachineAddress {
+        self.machine
+    }
+
+    pub fn pcode(&self) -> PcodeOffset {
+        self.pcode
+    }
+
+    pub(crate) fn add_pcode_offset(&self, off: PcodeOffset) -> Self {
+        Self {
+            machine: self.machine,
+            pcode: self.pcode.wrapping_add(off)
+        }
+    }
+    pub fn symbolize<'ctx>(&self, z3: &'ctx Context) -> SymbolicPcodeAddress<'ctx> {
+        SymbolicPcodeAddress(BV::concat(
+            &BV::from_u64(z3, self.machine, size_of::<PcodeMachineAddress>() as u32 * 8),
+            &BV::from_u64(z3, self.pcode as u64, size_of::<PcodeOffset>() as u32 * 8),
+        ))
+    }
+
+    pub fn resolve_from_varnode(vn: &VarNode, loc: ConcretePcodeAddress) -> Self {
+        if vn.is_const() {
+            // relative jump
+            loc.add_pcode_offset(vn.offset as u8)
+        } else {
+            // absolute jump
+            ConcretePcodeAddress {
+                machine: vn.offset,
+                pcode: 0
+            }
+        }
+    }
+}
+
+impl From<PcodeMachineAddress> for ConcretePcodeAddress {
+    fn from(value: PcodeMachineAddress) -> Self {
+        Self {
+            machine: value,
+            pcode: 0
+        }
+    }
+}
